Move necessary export headers
