--- src/bbox.h
@@ -1,1 +1,0 @@
-Rcpp::NumericVector CPL_get_bbox(Rcpp::List sf, int depth);
